Built-In Print Service
The Built-In Print Service implements a standard Android Print
Service and allows printing to Wi-Fi enabled printers on the users's
local network.

Printers must support mDNS discovery and the IPP protocol, and must
allow jobs to be sent in PCLm, PWG-raster, or PDF formats.

Internally, this service implements job control, page rendering, and
printer interactions in native code. This code relies on
libjpeg-turbo for JPEG encoding functions required by PCLm, and
selected portions of CUPS for direct interactions with printers.

Test: ./gradlew test
Change-Id: Ib2dca0190ba4abe028da1d9cd62308733ecce7b5
Signed-off-by: Mopria <mopriaplay@gmail.com>