make add printer dialog less invasive
When visiting the Add Printer activity for the first time with no
manual printers added, it makes sense to automatically prompt the user.
But on subsequent attempts the user may just want to see the list and
shouldn't have the dialog in the way.

With this change, the dialog only appears if the user has no manually-added
printers already.

Test: Use Add-Printer feature, won't show dialog if a printer already added
Change-Id: I5b9e38fa868ca0a4297a5dcc314d48d56fe7fee6
Signed-off-by: Glade Diviney <mopriadevteam@gmail.com>