Add support for Wi-Fi Direct printing
In environments where Wi-Fi LANs are inconvenient or unavailable, Wi-Fi Direct
connections allow users to print using an ad-hoc point-to-point network between
the mobile device and the printer. This code change allows users to discover
and add Wi-Fi Direct printers, and later select and print to them.

Because the same printer may appear both on a connected LAN and/or as a
an added Wi-Fi Direct printer, this code "de-duplicates" printer records based
on UUID. In this way, the same printer only appears once to the end user, even
if it is discovered using two different mechanisms.

The new package com.android.bips.p2p is added to manage Wi-Fi Direct discoveries
and connections. Other code is refactored as necessary according to current
style requirements. New UI Activities and Fragments are also added to expose
Wi-Fi Direct to the user.

Test: Add and print to a printer with Wi-Fi Direct capabilities
Change-Id: If4d8c7949ecc93851c07bd954674b33803b2fad5
Signed-off-by: Glade Diviney <mopriadevteam@gmail.com>