Support manual-add of IPPS and other printers
The current code does not allow IPPS-only printers to be added manually.
In addition, printers located at alternate URIs cannot currently be
added.

This changes addresses both problems by searching for additional ports/
paths when doing a manual add, and also by permitting the user to
specify a scheme (ipp:// or ipps://), a port number, and/or a path along
with the hostname. If any part is not supplied, known defaults are
populated and attempted. Any dangling network attempts are cancelled when
no longer needed.

Users can review the full printer URI in the details dialog of the Add
Printer activity.

Test: Set up an IPPS-only printer and try to add it manually
Change-Id: I34f77168d523a762b83adc692cbf6a444d831cd4
Signed-off-by: Glade Diviney <mopriadevteam@gmail.com>