Justify permission usage by built in print service
The location permission is only used as this app needs to scan the wifi
for printers on the same local network. In some cases we might store the
printer, but we are not storing the physical location of the printer.

The printer data will be backed up by the system the backup will not be
analyzed.

Test: Built
Change-Id: Iec17c3fa31ad030593c97d8be68bafd6646df0f6