Print photos in high-quality mode.
Printers may support draft, normal, and high-quality modes. The default
mode in this case is often "normal", with a balance between speed and quality.
But users printing photos are likely interested in a photo of the highest quality.
This change detects a printer with high-quality mode and automatically uses it
when printing a photo.

Test: Print a photo to a printer with different quality modes.
Change-Id: If7c70e89b66143ce28020bb37afaf0af65a77b87
Signed-off-by: Glade Diviney <mopriadevteam@gmail.com>