Allow deprecated declarations
In order to address the security vulnerability at b/111210196 we are
upgrading the libcups library. There are some calls to functions that
are deprecated between the versions; in order to minimize churn in
legacy branches we are allowing deprecated functions to be called rather
than modifying the code to use the new functions.

Bug: 111210196
Test: adb shell am instrument -w android.print.cts/androidx.test.runner.AndroidJUnitRunner
Change-Id: I2b7d568c9f8878d27d9e53f96930d62d07b7d84a