Set better share target for share-to-print
By adding the SUBSTITUTE_SHARE_TARGET_APP_NAME_AND_ICON the share
target's name can be replaced and instead of
"default print service - print" the share target is simply named
"print". As the default print service just sends to document to the main
printing API, this is the right behavior.

Test: Looked at share sheet
Fixes: 150646365
Change-Id: I2cf85d242682d774856a1939b43816d54ccf364d